Add files via upload
Implementation of a class that can be used to compute the solutions of linear and non-linear boundary value problems in one-dimension. The methods include a finite difference scheme for linear BVPs with Dirichlet conditions, a Newton-Raphson scheme for non-linear BVPs with Dirichlet conditions and the shooting method which can be used for both linear and non-linear BVPs with Dirichlet conditions. The shooting method also returns an estimate of the derivative of the solution of the BVP. The shooting method uses the IVP solver that was implemented previously. The finite difference method approximates a linear BVP by a system of linear equations, the Newton-Raphson method approximates the non-linear BVP by a system of non-linear equations.